+++ TomatoGame.Service/Dto/UserProfileDto.cs
@@ -17,6 +17,6 @@ namespace TomatoGame.Service.Dto
 
         public GameMode Mode { get; set; }
 
-        public float LatestScore { get; set; }
+        public double LatestScore { get; set; }
     }
 }
