+++ TomatoGame.Storage/Score.cs
@@ -14,7 +14,7 @@ namespace TomatoGame.Storage
 
         public int Mode { get; set; }
 
-        public float LatestScore { get; set; }
+        public double LatestScore { get; set; }
 
         public DateTime UpdatedTime { get; set; }
     }
